Move samp module out of pysamp module. Also fixed some incorrect size typing on race checkpoints Running strict on mypy - ensuring there are minimal amount of errors. Still need to fix and complete the textdraw module(s).

+++ samp/__init__.pyi
@@ -548,7 +548,7 @@ def DisableRemoteVehicleCollisions(player_id: int, disable: bool) -> bool:
 
 
 def SetPlayerCheckpoint(
-    player_id: int, x: float, y: float, z: float, size: int
+    player_id: int, x: float, y: float, z: float, size: float
 ) -> bool:
     pass
 
@@ -566,7 +566,7 @@ def SetPlayerRaceCheckpoint(
     next_x: float,
     next_y: float,
     next_z: float,
-    size: int,
+    size: float,
 ) -> bool:
     pass
